Frontend/MvcApp/CloudFunctions/Google_Cloud_Function/requirements.txt
flask==2.3.3
requests==2.31.0 
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.42.0
protobuf==4.25.0
pillow==10.1.0
numpy==1.26.4